chore: moves enums to separate files
