+++ packages/core/lib/core.dart
@@ -5,4 +5,5 @@ export './src/l10n/l10n.dart';
 export './src/models/models.dart';
 export './src/styles/styles.dart';
 export './src/utils/utils.dart';
+export './src/enums/enums.dart';
 export './src/widgets/widgets.dart';
