+++ packages/core/lib/src/enums/media_type.dart
@@ -1,0 +1,11 @@
+/// Enum for the type of media in a tweet
+enum MediaType {
+  /// The media is a video
+  video,
+
+  /// The media is a GIF
+  gif,
+
+  /// The media is a single image
+  photo,
+}
