+++ packages/core/lib/src/models/media.dart
@@ -1,14 +1,4 @@
-/// Enum for the type of media in a tweet
-enum MediaType {
-  /// The media is a video
-  video,
-
-  /// The media is a GIF
-  gif,
-
-  /// The media is a single image
-  photo,
-}
+import 'package:core/core.dart';
 
 /// Tweet Media object
 class Media {
